Modification des methode Admin pour la moderation en adequation avec le changement des ID

+++ Binouz/src/main/java/com/app/binouz/controller/CommentaireController.java
@@ -89,7 +89,7 @@ public class CommentaireController {
     
     
     @PostMapping("/deleteCom")
-    public String deleteBiereGet(@Valid Commentaire commentaire, BindingResult result,Model model, Integer idcom){
+    public String deleteBiereGet(@Valid Commentaire commentaire, BindingResult result,Model model, Long idcom){
